refactor: add service like methods to the service provider
